2.2
Variables function saving and storing data. Const is similar to stay an original state. So if you make a variable in JS. Use const;
And The way of making the name of a variable is camel case, snake case.
myName is camel case and my_name is snake case. camel case is usually used in JS. and snake case is usually used in Python. And lastly, What does console.log( ); means?
That means that print the data in '()' at the console.
So i want you know that log means print.