2.3
In the old version of JS, people use 'var' to use a variable. But the problem with 'var' is that programmers couldn't notice the updated point. So now we use 'const' or 'let' to use a variable. 'const' means the value should never be changed. As you can see, 'const' does not allow us to update. It allows us to only read. In contrast, 'let' means the value could be changed. Thanks to 'let', we can update the value freely. In conclusion, programmers can notice the purpose of each variable. So, you need to notice the purpose of a variable through 'let' and 'const' when you read code.