2.4 Booleans
true', 'false', 'null' is value.
But 'undefined' is not.
'undefined' means that a variable exists but the variable does not have any values. But the others exist and have values(true, false, null)
null means nothing. In one word, the value I don't know. And true and false means literally. Programmers use them to login, loading and some other things.