2.5 Arrays
Array functions to make a list.
In one word,  Array makes a group.
A programmer can add any values to the array. And you should remember that a computer count numbers from zero.  The first index of an array is zero. And you should remember if you want to add one more value, you can use "name.push". The value you want to add will be placed at the last index of an array.