2.6 Objects
Objects allow us to save data have some properties. So objects can store many data that have various
information. But in contrast, an array holds all the same meaning. so That is the difference between array and object. So I suggest you use 'array' if you want to make a shopping list.
And you had better use objects if you make the information of players.
By the way, we can also update and add one more value in 'const object' and 'const array'. Due to that fact, You can have a question "how can we update 'const'? yeah, correct. But you should know that we are trying to update the elements in the object and array. The browser can send us error messages if we try to const object and const array. But as I repeat, we have been trying to recode the elements in const object or const array. That's all.