2.7 2.8 functions part two
Nan means not a number.
And I study how to get the data inside of the function. Through the variable of function, the function gets the data from outside. And the variable of the function only can exist in the function. Outside of the function, the values have not existed.
And you can also declare a function in the object. That way of making functions is a little bit different.
Inside of the box, you can declare a function like sayHello:function(var){};
Outside of the box, you can declare a   function like function sayHello(){};