2.11 2.12 Return
See! Why do we need to use 'return'?
If I use only 'console.log' in this program, I couldn't calculate the sum of quotient and remainder. I have to get the value from the function. And then I calculate the sum of the quotient and remainder.
So all I have to know is that 'return' is similar to the mixer. We put many ingredients into the mixer. Like
"mixer(apple,milk)".  The mixer can show the result of its work and we can see. Like console.log(apple+milk).
But sometimes we want to mix something to the result of the mixer.
That's when I should get out the result from inside of the mixer.
And then I ask how?
The answer is 'pour';
The mixer 'pours' the result into a 'plate'. I mean pouring is 'return' and the plate is 'variable'. Okay, welllll
In conclusion,
mixer = function
pour = return
plate = variable