2.13 conditionals part one
The prompt function pops up the message to a user and gets the value from a user. And the type of value is always a string. Use 'typeof' if you need to check the type of value.
Anyway, the value from the 'prompt' function is always a string. So If you toss the number to the function 'prompt', the function will understand the value is a string. But you don't want. That's when I use the 'parseInt' function. Through the function, you can change the type of the value to an integer.