2.14 conditionals part Two
If the function 'parseInt' get a string as an input value, It will reply 'Nan'.

And the function 'isNaN' means "Is that not a number?".
So if the function gets an integer, It will return 'False'. But if the function gets a value not an integer, it will return 'true'.
And now we have to know "If and else". If the condition is true, the flow should go toward the inside of 'if'.
But if the condition is false, the flow should go toward the inside of 'else'.
And you need to know the 'else' is not essential when you use 'if'.