2.15 2.16 conditional part Three & Recap
If the condition is false, the conditional should keep going to find true.
If the condition is true, the conditional should
execute the order in the 'else if' or 'if' and break outside of the conditional. And If all the conditions were false,
The flow of conditional will go to the 'else'. In the 'else', a computer executes the orders in the 'else' and breaks outside of the conditional.

Finally, I learned some operators.
'=' => assign a value
'===' => check is the same?
'&&' => if all the values were true, it will return true.
'||' => if one of the values was true, it will return true.