3.3 events
function handleTitleClick ( ) {
    title.style.color ="blue";
}
title.addEventListener("click", handleTitleClick);

This code shows well the effect of addEventListener.

addEventListener is going to try to hear the events you added.
In this case, the event you added is 'click'. If one user 'clicks', addEventListener will catch the event 'click'. And then addEventListern will execute the react you wrote.
In this case, the react you wrote is the function 'handleTitleClick'.