3.5 More Evetns
There are two ways of catching events.
The first one is addEventListener.
And the other one is oneventname.
h1.onEventName =functionName;
The lecturer recommends using addEventListener. Because we can use
'h1.removeEventListener;' together with addEventListener.

And The 'window' represents a window containing a DOM document. The 'window' already exists like 'document'. we can use the objects through JavaScript