3.6 CSS in Javascript
The first point is the effect of variables.
It is comfortable to understand intentions.

The second point is to use conditionals. It is really easy to express various reactions.

And the most important point we have to concentrate on is the process.

We can bring the element that I want through querySeletor.
And then we can realize the event related to the element through addEventListener or onEventName.

And then we can order some reactions through functions when the event breaks out.

In addition, you can use conditionals to represent various reactions.