3.7 CSS in Javascript part Two
The things related to styles had better written in the CSS file. I mean coding the things related to style in CSS files is a better way.

I changed the 'className' of the h1 tags in this program with JavaScript. The process is a domino effect.
JavaScript changed the 'className' of the h1 tag in HTML. In short, JavaScript changed the HTML file. And then HTML file refers to a CSS file.

In this program, two errors were predicted.
The first expected error is 'string'. If you use a string in the conditions, you might make a mistake. Because misspelling is a natural error way. The solution to the misspelled error is to use a constant variable. In one word, save a string to a constant variable. This method will shield misspelled errors.
The second expected error is related to the original className. In the circumstance that I change the className in JavaScript, the original className of the h1 tag disappeared.
So When you change className,  you should care about the original className.