3.8 CSS in Javascript part Three (1)
The method 'classList' means literally.
I mean if I treat the class name through 'className', the method will replace everything.  However, If I use 'classList', The method can allow us to handle class an array. So we can treat the class as each element through 'classList'. And we can use also functions if I use 'classList'.

"element.classList.contains(String)" can check the existence and nonexistence of the class on the element. The result from the function is True or False.

"element.classList.add(string)"
can add a class to the element.
If the class that was added to classList exists, the code will be ignored.

"element.classList.remove(String)"
can delete a class from the element.
If the class is nonexistent in classList, This code will be ignored.

"element.classList.toggle(String)"
can add a class if the class that was added to classList doesn't exist in the
classList. If the circumstance is in contrast to the case I said, the action will be reversed.