3.8 CSS in Javascript part Three (2)
You can see how 'toggle' is powerful.
The function 'toggle' replace the lines that include 'contains', 'remove', 'add'.

As I said, the toggle can add a class to classList If a class that was added doesn't exist in the classList.

And the toggle also can delete a class in the classList.
The case is that the value given by the programmer is already existent in the classList.