4.0  Input Values
You need to check the way of accessing the id. 'getElementById' access the id with '#'. In contrast, 'querySelector' don't need to any character.

In this program, I learned how to get the data when a user click the log-in button.
addEventListener and value are needed to get the data from the click.
The value attribute specifies the value of an <input> element.