4.1 Form Submission (2)
Coding the sheet of javascript is a good way. But sometimes we need to use the things provided by the browser. I mean, We need to use HTML effectively. We can check the input is blank and the max length of input through 'form'. In short, we need to use the help of the browser.

And as you remember, the tag 'input' should be in the tag 'form' if you want to check validation.

And the problem with this program
is re-start when you submit or click the button in the 'form'. I mean,
the page is restarted when you submit or click the button in the form.