4.2 Events
The function preventDefault() block the default behavior by the browser.
The function is effective for 'a', 'submit', and 'button'.
In this program, I used the function 'preventDefault' to block re-start which is the default behavior.

And the first argument of all in the function "addEventListener" will be information about what is happened.
JavaScript gives the information for free.
So in this case, the argument 'event' will be the information about 'submit'.