4.3 Events part Two
"link.addEventListener("click",onLink);" have the function 'onlink'.
The function 'onlink' is executed by JavaScript. I mean, you did not execute the function. The function is executed by Javascript.

Why do we need to use the function 'preventDefault()'? Because sometimes we want to get the info about what happened. As stopping the default behavior, we can check the information about what happened. For example, I want to know the information when I click the link. But the information about the clicking link event goes away as the page move to the other page according to the hyperlink. So we need to use the function "preventDefault()". Okay. Well done.

In conclusion, the function inside of the 'addEventListener' is not executed directly by the user. The browser
that has JavaScript executes the function in the addEventListener.
In addition, The browser will give the information about the event. All we have to do for getting the information is to make the argument
in the function.