4.4 Getting Username
In this program, I used the function 'add' and 'remove' with 'classList'.
And I also assign a string using 'innerText'.

String concatenation has two different ways. The first way is     /*greeting.innerText = "hello " + username;*/ But this way is a little bit uncomfortable. Because I should give "hello" the space.

So the lecture recommends that using backticks is better.
The way looks like this, "greeting.innerText = `hello ${username}`;"

You should know the two cautions.
The first one is backtick, ' ` '.
And put the variable to the space,
${ }.