4.6 Loading Username
The string often makes the developer misspell. So please store the string to the upper case letter variable.

In this program, I wanted to make the browser remember the username.
And then I wanted to remove the 'form' after that I entered the username. Also, I wanted to show the header tag that shows the username after that I entered the username.

One of the purposes of the 'IF' conditionals is to show the form.
And the other purpose is to call 'addEvenListener'.

By calling the function 'onSubmit' by 'addEvenListener', the 'form' was deleted. and the value of the input is stored in the local storage. The key is 'username' and the value is the value of the input.
And lastly, the function 'onSubmit' calls the function 'paintGreetings'.

The function 'paintGreetings' assign the input that the user entered to the header that has 'greeting'.
And then by deleting the class 'hidden', the browser can show the tag on the screen.

And the program meets the 'else' conditional at the end of the program. In the 'else' conditionals', the program will call the function 'paintGreetings'.
By executing the function 'paintGreetings', we can store the username and show the username on the screen if we restart the page.