7.0 Setup 7.1 Adding ToDos
First of all, if you make a change to a file, you should think about the route of the file. Like "js/greetings.js", "css/style.css".

Let me explain the file 'todo.js'.

'toDoform' get the form from HTML.
'toDoInput' get the input of the form.
'toDoList; get the list.

And If you submit something, the "toDoForm.addEventListener" will catch the submit and will call the function "handleToDoSubmit".

Through the "handleToDoSubmit",
the default behavior of submission is blocked by "event.preventDefault();"
And the variable "newTodo" save the toDoInput.value that you enter.
Also, The input will be a empty space  by the code "toDoInput.value = "";".
And the last action of the function
"handleToDoSubmit" is to call the function "paintToDo".

As you can see, the function "paintToDo" get the argument "newTodo" from the function "handleToDoSubmit".

And the argument will be assigned to the variable 'newTodo' in the function
"paintToDo". Maybe you can be confused because the name is the same. So You can make another name if you want.

And using the function "paintToDo",
we made 'li' and 'span' in the HTML.

And the span was added to li by the code  "li.appendChild(span);".
And then, the span gets the value of the variable "newTodo" by the code
"span.innerText = newTodo;".
In short, the result of the function "paintToDo" is to make the list element that has a span that gets the "newTodo". And the last action of the function "paintToDo" is to assign the list element to the inside of "ul".

If you enter something into the input,
the action I explained will be executed repeatedly.