7.3 Saving to Dos
The code "toDos.push(newTodo);" can insert a value to the array.

The code "localStorage.setItem" can store a value at the storage.

The code "JSON.stringify" make all the things string.

All the things in the local storage are text(string). So if we want to store an array in the local storage, we should convert the array to a string.
That's when we need to use "JSON.stringify(toDos)".